fix: races page functionality after Jotai refactor
- Convert races page to Server/Client Component hybrid pattern for proper auth
- Fix API response format (return array directly instead of object)
- Create asyncRacesAtom with proper data fetching and refresh capability
- Fix race creation, display, and training plan attachment functionality
- Remove problematic test files that were causing CI issues
