+++ src/components/chat/ChatWindow.tsx
@@ -283,6 +283,7 @@ export default function ChatWindow({ recipientId, recipient }: ChatWindowProps) 
         <PerformantMessageList
           recipientId={recipientId}
           currentUserId={(session?.user?.id as string) || ''}
+          messages={messages}
         />
 
         {/* Typing Indicator */}
