--- tests/manual-auth-test.manual.ts
@@ -1,76 +1,0 @@
-import { expect, test } from '@playwright/test'
-
-test.describe('Manual Authentication Test', () => {
-  test('should login successfully and redirect to correct dashboard', async ({ page }) => {
-    // Navigate to signin page
-    await page.goto('/auth/signin')
-
-    // Wait for page to load completely (removed networkidle - causes CI hangs)
-    await page.waitForSelector('input[type="email"]', { state: 'visible' })
-
-    console.log('🧪 Testing login with alex.rivera@ultracoach.dev')
-
-    // Fill email
-    await page.fill('input[type="email"]', 'alex.rivera@ultracoach.dev')
-
-    // Fill password
-    await page.fill('input[type="password"]', 'RunnerPass2025!')
-
-    // Submit form and wait for navigation
-    await Promise.all([
-      page.waitForURL(/dashboard\/runner/, { timeout: 15000 }),
-      page.click('button[type="submit"]'),
-    ])
-
-    console.log('✅ Successfully logged in and redirected to runner dashboard')
-
-    // Verify we're on the runner dashboard
-    await expect(page).toHaveURL(/dashboard\/runner/)
-
-    // Take a screenshot for verification
-    await page.screenshot({ path: 'successful-runner-login.png' })
-  })
-
-  test('should login coach successfully and redirect to coach dashboard', async ({ page }) => {
-    // Navigate to signin page
-    await page.goto('/auth/signin')
-
-    // Wait for page to load completely (removed networkidle - causes CI hangs)
-    await page.waitForSelector('input[type="email"]', { state: 'visible' })
-
-    console.log('🧪 Testing login with emma@ultracoach.dev')
-
-    // Fill email
-    await page.fill('input[type="email"]', 'emma@ultracoach.dev')
-
-    // Fill password
-    await page.fill('input[type="password"]', 'UltraCoach2025!')
-
-    // Submit form
-    await page.click('button[type="submit"]')
-
-    // Wait for navigation with more flexible pattern
-    await page.waitForURL(
-      url => {
-        const path = new URL(url).pathname
-        return path.includes('/dashboard/coach') || path === '/' || path.includes('/dashboard')
-      },
-      { timeout: 30000 }
-    )
-
-    // If we landed on home page, navigate to coach dashboard
-    const currentUrl = page.url()
-    if (currentUrl.endsWith('/') || !currentUrl.includes('/dashboard/coach')) {
-      await page.goto('/dashboard/coach')
-      await page.waitForURL(/dashboard\/coach/, { timeout: 15000 })
-    }
-
-    console.log('✅ Successfully logged in and redirected to coach dashboard')
-
-    // Verify we're on the coach dashboard
-    await expect(page).toHaveURL(/dashboard\/coach/)
-
-    // Take a screenshot for verification
-    await page.screenshot({ path: 'successful-coach-login.png' })
-  })
-})
